--- service/user/src/main/java/com/sparta/user/domain/model/vo/UserRole.java
@@ -1,8 +1,0 @@
-package com.sparta.user.domain.model.vo;
-
-public enum UserRole {
-    MASTER,
-    HUB_MANAGER,
-    HUB_SHIPPING_MANAGER,
-    HUB_COMPANY
-}
