+++ service/user/user_dto/src/main/java/com/sparta/user/dto/user_dto/UserCreateDto.java
@@ -1,0 +1,16 @@
+package com.sparta.user.dto.user_dto;
+
+/**
+ * 사용자 생성시 요청할 dto
+ *
+ * @param username
+ * @param email
+ * @param encodedPassword
+ */
+public record UserCreateDto(
+    String username,
+    String email,
+    String encodedPassword
+) {
+
+}
