+++ service/user/user_dto/src/test/java/com/sparta/user/dto/user_dto/UserDtoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sparta.user.dto.user_dto;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserDtoApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
