+++ service/user/user_server/src/main/java/com/sparta/user/application/service/UserService.java
@@ -6,4 +6,5 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 @Service
 public class UserService {
+
 }
