+++ service/user/user_server/src/main/java/com/sparta/user/domain/model/vo/ShippingManagerType.java
@@ -1,6 +1,6 @@
 package com.sparta.user.domain.model.vo;
 
 public enum ShippingManagerType {
-    HUB_MOVING_AGENT,
-    COMPANY_DELIVERY_AGENT
+  HUB_MOVING_AGENT,
+  COMPANY_DELIVERY_AGENT
 }
