+++ service/user/user_server/src/main/java/com/sparta/user/domain/model/vo/UserRole.java
@@ -1,0 +1,8 @@
+package com.sparta.user.domain.model.vo;
+
+public enum UserRole {
+  MASTER,
+  HUB_MANAGER,
+  HUB_SHIPPING_MANAGER,
+  HUB_COMPANY
+}
