+++ service/user/user_server/src/main/java/com/sparta/user/presentation/controller/UserController.java
@@ -10,5 +10,5 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/users")
 public class UserController {
 
-    private final UserService userService;
+  private final UserService userService;
 }
