+++ service/user/user_server/src/main/resources/application.yml
@@ -3,7 +3,7 @@ server:
 
 spring:
   application:
-    name: user
+    name: user-service
 
   datasource:
     driver-class-name: org.postgresql.Driver
